add ft_realloc function and refactor expander
