+++ Libft/srcs_libft/ft_realloc.c
@@ -1,0 +1,38 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_realloc.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jgomes-v <jgomes-v@student.42porto.com>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/22 17:42:52 by jgomes-v          #+#    #+#             */
+/*   Updated: 2023/11/22 17:43:23 by jgomes-v         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/libft.h"
+
+void	*ft_realloc(void *ptr, size_t new_size)
+{
+	void	*new_ptr;
+
+	if (new_size == 0)
+	{
+		free(ptr);
+		return (NULL);
+	}
+	new_ptr = malloc(new_size);
+	if (new_ptr)
+	{
+		if (ptr)
+		{
+			ft_memcpy(new_ptr, ptr, new_size);
+			free(ptr);
+		}
+		return (new_ptr);
+	}
+	else
+	{
+		return (NULL);
+	}
+}
